Extract handler for SendEmail events
